Revert "housekeeping: refactor into global and local component from main page"
