+++ apps/web/app/[id]/constans.ts
@@ -1,5 +1,4 @@
 import * as yup from 'yup';
-import { FormEntry } from './types';
 
 export const editSchedulingSchema = () =>
   yup.object().shape({
@@ -7,18 +6,3 @@ export const editSchedulingSchema = () =>
     brand: yup.string().typeError('must be string'),
     barcode: yup.number().typeError('must be number'),
   });
-
-export const formList: FormEntry[] = [
-  {
-    name: 'name',
-    label: 'Product Name',
-  },
-  {
-    name: 'brand',
-    label: 'Brand Name',
-  },
-  {
-    name: 'barcode',
-    label: 'Barcode',
-  },
-];
