--- apps/web/app/[id]/types.ts
@@ -1,9 +1,0 @@
-export interface FormEntry {
-  name: string;
-  label: string;
-}
-
-export interface NotificationType {
-  message: string;
-  severity: 'success' | 'warning';
-}
