+++ apps/web/app/[id]/useProductStore.ts
@@ -1,7 +1,6 @@
 import { useState } from 'react';
 import { APIRemote, Product } from '../types';
 import axios from 'axios';
-import { NotificationType } from './types';
 
 const useProductStore = () => {
   const [product, setProducts] = useState<Product>({
@@ -13,26 +12,7 @@ const useProductStore = () => {
   });
   const [error, setError] = useState<string | undefined>(undefined);
   const [loading, isLoading] = useState<boolean>(true);
-  const [isNotify, toggleIsNotify] = useState<boolean>(false);
-  const [notifyData, setNotifyData] = useState<NotificationType>({
-    message: '',
-    severity: 'success',
-  });
 
-  const openNotification = () => {
-    toggleIsNotify(true);
-  };
-
-  const handleClose = (
-    _event: React.SyntheticEvent | Event,
-    reason?: string,
-  ) => {
-    if (reason === 'clickaway') {
-      return;
-    }
-
-    toggleIsNotify(false);
-  };
   const fetchProduct = async (id: string) => {
     try {
       const url = 'http://localhost:3333/products/' + id;
@@ -47,36 +27,21 @@ const useProductStore = () => {
     }
   };
 
-  const updateProduct = async (product: Product) => {
+  const updateProduct = async (
+    product: Product,
+    callback: (success: boolean) => void,
+  ) => {
     try {
       const url = 'http://localhost:3333/products';
       const { data } = await axios.put<APIRemote<Product>>(url, { ...product });
-      data.data &&
-        setNotifyData({
-          severity: 'success',
-          message: 'Saved!',
-        });
+      data.data && callback(true);
     } catch (error) {
-      setNotifyData({
-        severity: 'warning',
-        message: 'Error on save!',
-      });
+      callback(false);
 
       console.log(error);
     }
-    openNotification();
   };
-  return {
-    product,
-    loading,
-    error,
-    isNotify,
-    notifyData,
-    openNotification,
-    handleClose,
-    fetchProduct,
-    updateProduct,
-  };
+  return { product, loading, error, fetchProduct, updateProduct };
 };
 
 export default useProductStore;
