--- apps/web/components/index.tsx
@@ -1,4 +1,0 @@
-//This is global component
-
-export { default as SearchBar } from './SearchBar';
-export { default as MediaCard } from './MediaCard';
